WIP - Update library, remove dead code, update pre-commit, use pyproject.toml, use src layout,
