+++ .github/workflows/main.yml
@@ -1,15 +1,18 @@
 name: build
 
-on: [ push, pull_request ]
+on:
+  push:
+    branches:
+      - main
+  pull_request:
 
 jobs:
   build:
-
     name: Conda Build with Python${{ matrix.python-version }}
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: ["3.9", "3.10", "3.11"]
+        python-version: ["3.9", "3.10", "3.11", "3.12"]
     defaults:
       run:
         shell: bash -l {0}
@@ -26,16 +29,10 @@ jobs:
         cache-environment: true
         environment-file: environment.yml
         create-args: >-
-          conda
           python=${{ matrix.python-version }}
-    - name: Conda and Mamba versions
-      run: |
-        conda --version
-        echo "micromamba: $(micromamba --version)"
     - name: Install dependencies
       run: |
-        python -m pip install --no-user --editable "."
-        if [ -f requirements_dev.txt ]; then pip install -r requirements_dev.txt; fi
+        python -m pip install --no-user --editable ".[dev]"
     - name: Lint with flake8
       run: make lint
       if: matrix.python-version == 3.9
