--- MANIFEST.in
@@ -1,9 +1,0 @@
-graft daops
-include *.md
-global-exclude __pycache__
-global-exclude *.py[co]
-
-recursive-include etc *
-include LICENSE
-include requirements.txt
-include requirements_dev.txt
