+++ README.rst
@@ -4,21 +4,20 @@ daops - data-aware operations
 
 
 .. image:: https://img.shields.io/pypi/v/daops.svg
-   :target: https://pypi.python.org/pypi/daops
-   :alt: Pypi
-
-
+    :target: https://pypi.python.org/pypi/daops
+    :alt: PyPI
 
 .. image:: https://github.com/roocs/daops/workflows/build/badge.svg
     :target: https://github.com/roocs/daops/actions
     :alt: Build Status
 
-
-
 .. image:: https://readthedocs.org/projects/daops/badge/?version=latest
-   :target: https://daops.readthedocs.io/en/latest/?badge=latest
-   :alt: Documentation
+    :target: https://daops.readthedocs.io/en/latest/?badge=latest
+    :alt: Documentation
 
+.. image:: https://img.shields.io/badge/code%20style-black-000000.svg
+    :target: https://github.com/python/black
+    :alt: Python Black
 
 The ``daops`` library (pronounced "day-ops") provides a python interface to a
 set of operations suitable for working with climate simulation outputs. It is
@@ -32,7 +31,6 @@ The data will be loaded and *fixed* using the `xarray <http://xarray.pydata.org/
 library before the any actual operations are sent to its sister library
 `clisops <https://github.com/roocs/clisops>`_.
 
-
 * Free software: BSD
 * Documentation: https://daops.readthedocs.io
 
@@ -40,7 +38,6 @@ Features
 --------
 
 The package has the following features:
-
 
 * Ability to run *data-reduction* operations on large climate data sets.
 * Knowledge of irregularities/anomalies in some climate data sets.
@@ -52,11 +49,5 @@ Credits
 
 This package was created with ``Cookiecutter`` and the ``cedadev/cookiecutter-pypackage`` project template.
 
-
 * Cookiecutter: https://github.com/audreyr/cookiecutter
 * cookiecutter-pypackage: https://github.com/cedadev/cookiecutter-pypackage
-
-
-.. image:: https://img.shields.io/badge/code%20style-black-000000.svg
-   :target: https://github.com/python/black
-   :alt: Python Black
