+++ environment.yml
@@ -5,15 +5,16 @@ channels:
   - defaults
 dependencies:
   - pip
-  - numpy >=1.16
-  - xarray >=0.21,<2023.3.0 # https://github.com/pydata/xarray/issues/7794
+  - cftime
+  - clisops >=0.14.0
   - dask
+  - elasticsearch>=8.0.1
   - netcdf4
-  - clisops >=0.12.1
+  - numpy >=1.23.0
   - roocs-grids >=0.1.2
+  - roocs-utils >=0.6.4
+  - xarray >=0.21,<2023.3.0 # https://github.com/pydata/xarray/issues/7794
   - xesmf >=0.8.2
-  - elasticsearch>=7.9.1
-  - roocs-utils>=0.6.2
   # logging
   - loguru >=0.5.3
   # catalog
