--- requirements.txt
@@ -1,19 +1,0 @@
-numpy>=1.16
-xarray>=0.21
-xesmf>=0.8.2
-dask[complete]
-cftime
-netcdf4
-elasticsearch>=8.0.1
-clisops>=0.12.1
-## clisops @ git+https://github.com/roocs/clisops.git@master#egg=clisops
-roocs-utils>=0.6.4
-roocs_grids>=0.1.2
-# logging
-loguru>=0.5.3
-# catalog
-intake>=0.7.0,<2.0
-# to support kerchunk
-aiohttp
-fsspec
-zarr==2.13.3
