--- requirements_dev.txt
@@ -1,19 +1,0 @@
-pip
-bumpversion
-wheel
-watchdog
-flake8
-tox>=4.0
-coverage
-Sphinx
-sphinx-rtd-theme
-nbsphinx
-nbconvert
-ipython
-twine
-pytest>=6.0
-pytest-loguru
-pytest-runner
-pre-commit>=3.0.0
-black>=23.7.0
-GitPython>=3.1.37
