+++ roocs.ini.tmpl
@@ -1,17 +1,17 @@
 [project:cmip5]
-base_dir = DATA_DIR/test_data/badc/cmip5/data/cmip5
+base_dir = DATA_DIR/badc/cmip5/data/cmip5
 
 [project:cmip6]
-base_dir = DATA_DIR/test_data/badc/cmip6/data/CMIP6
+base_dir = DATA_DIR/badc/cmip6/data/CMIP6
 
 [project:cordex]
-base_dir = DATA_DIR/test_data/badc/cordex/data/cordex
+base_dir = DATA_DIR/badc/cordex/data/cordex
 
 [project:c3s-cmip5]
-base_dir = DATA_DIR/test_data/gws/nopw/j04/cp4cds1_vol1/data/c3s-cmip5
+base_dir = DATA_DIR/gws/nopw/j04/cp4cds1_vol1/data/c3s-cmip5
 
 [project:c3s-cmip6]
-base_dir = DATA_DIR/test_data/badc/cmip6/data/CMIP6
+base_dir = DATA_DIR/badc/cmip6/data/CMIP6
 
 [project:c3s-cordex]
-base_dir = DATA_DIR/test_data/gws/nopw/j04/cp4cds1_vol1/data/c3s-cordex
+base_dir = DATA_DIR/gws/nopw/j04/cp4cds1_vol1/data/c3s-cordex
