+++ src/daops/catalog/base.py
