+++ src/daops/catalog/util.py
