+++ src/daops/data_utils/__init__.py
@@ -1,0 +1,6 @@
+from .coord_utils import squeeze_dims
+
+
+# Demonstrator dummy function
+def do_nothing(ds):
+    return ds
