+++ src/daops/data_utils/attr_utils.py
@@ -1,5 +1,3 @@
-from roocs_utils.xarray_utils import xarray_utils as xu
-
 from .common_utils import handle_derive_str
