+++ src/daops/ops/subset.py
@@ -1,7 +1,6 @@
 from clisops.ops.subset import subset as clisops_subset
-from roocs_utils.parameter import parameterise
-
 from daops.ops.base import Operation
+from roocs_utils.parameter import parameterise
 
 __all__ = [
     "subset",
