+++ src/daops/processor.py
