+++ src/daops/utils/common.py
