+++ tests/test_data_utils/test_attr_utils.py
@@ -1,10 +1,10 @@
 import numpy as np
 import xarray as xr
-from roocs_utils.xarray_utils.xarray_utils import open_xr_dataset
-
 from daops.data_utils.attr_utils import add_global_attrs_if_needed
 from daops.data_utils.attr_utils import edit_global_attrs
 from daops.data_utils.attr_utils import edit_var_attrs
+from roocs_utils.xarray_utils.xarray_utils import open_xr_dataset
+
 from tests._common import CMIP6_DECADAL
 from tests._common import MINI_ESGF_MASTER_DIR
