+++ tests/test_data_utils/test_coord_utils.py
@@ -1,11 +1,11 @@
 import cftime
 import numpy as np
 import xarray as xr
-from roocs_utils.xarray_utils.xarray_utils import open_xr_dataset
-
 from daops.data_utils.coord_utils import add_coord
 from daops.data_utils.coord_utils import add_scalar_coord
 from daops.data_utils.coord_utils import squeeze_dims
+from roocs_utils.xarray_utils.xarray_utils import open_xr_dataset
+
 from tests._common import CMIP6_DECADAL
 from tests._common import MINI_ESGF_MASTER_DIR
