+++ tests/test_data_utils/test_var_utils.py
@@ -1,6 +1,6 @@
 import xarray as xr
-
 from daops.data_utils.var_utils import add_data_var
+
 from tests._common import MINI_ESGF_MASTER_DIR
