+++ tests/test_fixes_applied.py
@@ -5,9 +5,9 @@ import cftime
 import numpy as np
 import pytest
 import xarray as xr
-
 from daops import CONFIG
 from daops.ops.subset import subset
+
 from tests._common import MINI_ESGF_MASTER_DIR
