+++ tests/test_logging.py
@@ -1,9 +1,9 @@
 import sys
 
 import pytest
-
 from daops.utils.common import _logging_examples
 from daops.utils.common import enable_logging
+
 from tests._common import ContextLogger
