+++ tests/test_utils/test_normalise.py
@@ -1,6 +1,7 @@
 from collections import OrderedDict
 
 from daops.utils.normalise import ResultSet
+
 from tests._common import MINI_ESGF_MASTER_DIR
