+++ tests/test_xarray/test_xarray_aggregation.py
@@ -12,6 +12,7 @@ These tests use 3 NetCDF files (reduced sized), referenced by variables:
  - F3
 
 """
+
 import itertools
 import os
 import pathlib
