+++ tox.ini
@@ -10,7 +10,6 @@ python =
     3.7: docs
     3.7: black
 
-
 [testenv:black]
 skip_install = True
 basepython = python
